requirements.txt
# OpenAI with Realtime API support
openai>=1.50.0

# Web framework for deployment
flask>=3.0.0
flask-socketio>=5.3.0

# WebRTC and Real-time communication
websockets>=12.0
aiohttp>=3.9.0

# Audio processing for WebRTC streaming
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Core dependencies
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0

# Document processing
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
reportlab>=4.0.0

# Additional utilities
aiofiles>=23.0.0
structlog>=23.2.0
click>=8.1.0
